Change rating and stddev to decimal

+++ Core/EntityStores/Contracts/Models/PlayerEntity.cs
@@ -9,8 +9,8 @@ public class PlayerEntity : BaseEntity
     public required long SteamId { get; set; }
     public required string SteamUsername { get; set; }
     public required DateTime UsernameLastUpdatedDatetime { get; set; }
-    public double Rating { get; set; } = 1000;
-    public double Deviation { get; set; } = 1000 / 3;
+    public decimal Rating { get; set; } = 1000;
+    public decimal Deviation { get; set; } = 1000 / 3;
     public ICollection<KillEventEntity> Kills { get; set; } = new List<KillEventEntity>();
     public ICollection<KillEventEntity> Deaths { get; set; } = new List<KillEventEntity>();
     public ICollection<PlayerMatchStatEntity> PlayerMatchStats { get; set; } = new List<PlayerMatchStatEntity>();
